+++ Requirements2Z3/src/main/java/requirements2Z3/factory/BeUfFsFactory.java
@@ -8,7 +8,7 @@ import requirements2Z3.visitors.translators.BeUfFs;
 
 public class BeUfFsFactory extends EncoderFactory {
 
-	public BeUfFsFactory(int bound, float timestamp) {
+	public BeUfFsFactory(int bound, double timestamp) {
 		super(new BeUfFs(new BoundedEncoder(new FixedStepEncoder(new UfBasedEncoding(), timestamp), bound), bound, timestamp));
 	}
 }
